Fix - move authProvider to utils/ folder, fix checkLocalToken function in app store naming to checkLocalTokenIsValid
