+++ src/components/Header.js
@@ -4,7 +4,7 @@ import { backgroundColor1 } from "styles/colors"
 import { ReactComponent as MediumLogo } from "assets/MediumLogo.svg"
 import { ReactComponent as SmallLogo } from "assets/SmallLogo.svg"
 import { useMediaQuery } from "react-responsive"
-import { logout, checkLocalToken } from "authProvider"
+import { logout, checkLocalToken } from "utils/authProvider"
 import Dropdown from "components/Dropdown"
 import { body2Normal } from "styles/textTheme"
 import { Link, withRouter } from "react-router-dom"
